mas install status, so will not reinstall existing ones

+++ mas/appstore.sh
@@ -10,7 +10,12 @@ if type "mas" > /dev/null; then
   mas upgrade
   echo "  Installing App Store apps."
   while read p; do
-    mas install $p
+    if [ ! -f $ZSH/mas/$p ]; then
+        mas install $p
+        touch $ZSH/mas/$p
+    else
+        echo "Skipping "$p
+    fi
   done < $ZSH/mas/applications.list
 else
   echo "  Did not find mas, skipping app store applications."
